requirements.txt
openai
gtts
speechrecognition
pyaudio
python-dotenv
langdetect